scale the code in component and utils
